Actualizacion de model, repository y controller

+++ backend/src/main/java/com/grindsup/backend/repository/EntrenadorRepository.java
@@ -3,6 +3,5 @@ package com.grindsup.backend.repository;
 import com.grindsup.backend.model.Entrenador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntrenadorRepository extends JpaRepository<Entrenador, Long> {
-
+public interface EntrenadorRepository extends JpaRepository<Entrenador, Integer> {
 }
